EOM codes (EE,EA,IP,DIP,DEA) tested to make sure that the subspace truncation logic is correct.
Also checked to see whether old subspace can be augmented for the
next root (instead of starting over).
For DIP and DEA there still seems to be bug (biorthogonalization does
not come out right). Need to revist this and initial guess for DIA
and DIP.

Some modifications for fractional occupation code.